add GetConfirmedBlocks
